+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.9
+Bug fixes
+* flutter retrieve all plans issue - Response has no/invalid body. And the issue fixed in Native iOS SDK v1.0.17
+* Added configure() method callbacks to handle error case if any errors thrown from Chargebee during configuration
+  Fixed in this PR - https://github.com/chargebee/chargebee-ios/pull/56 and https://github.com/chargebee/chargebee-ios/pull/55
 ## 0.0.8
 Bug fixes
 * Android app crashed on signed apk and the issue fixed in Native Android SDK v1.0.15